requirements.txt
fastapi                 
uvicorn     
python-multipart        

groq                   
streamlit

sqlalchemy              
alembic
asyncpg                 
psycopg2-binary         
          
pydantic
pydantic[email]                
pydantic-settings   
python-dotenv   
pytest
httpx
pytest-asyncio
requests               